// src/interfaces/customs/sarlaft/QueriesOwnList.ts
export interface ISarlaftQueriesOwnList {
  id: number
  type_identification: string
  identification_number: string
  name: string
  list: string
  'registration date': string
  last_page: number
}

export interface ISarlaftAllOwnList {
  id: number
  name: string
  status: string
  created_at: string
  updated_at: string
}